+++ src/cljs/matthiasn/systems_toolbox/reagent.cljs
@@ -6,9 +6,8 @@
 (defn init
   "Return clean initial component state atom."
   [reagent-cmp-map dom-id init-state init-fn put-fn]
-  (let [local (if init-state
-                (atom init-state)
-                (atom {}))
+  (let [initial-state (or init-state {})
+        local (atom initial-state)
         observed (atom {})
         cmd (fn ([& r] (fn [e] (.stopPropagation e) (put-fn (vec r)))))
         reagent-cmp (create-class reagent-cmp-map)
@@ -19,10 +18,12 @@
     (r/render-component [reagent-cmp view-cmp-map] (by-id dom-id))
     (when init-fn (init-fn view-cmp-map))
     {:state {:local local
-             :observed observed}}))
+             :observed observed
+             :initial-state initial-state}}))
 
-(defn state-pub-handler
-  "Handle incoming messages: process / add to application state."
+(defn default-state-pub-handler
+  "Default handler function, can be replaced by a more application-specific handler function, for example
+  for resetting local component state when user is not logged in."
   [{:keys [cmp-state msg-payload]}]
   (reset! (:observed cmp-state) msg-payload))
 
@@ -30,14 +31,14 @@
   "Creates a component map for a UI component using Reagent. This map can then be used by the comp/make-component
   function to initialize a component. Typically, this would be done by the switchboard."
   {:added "0.3.1"}
-  [{:keys [cmp-id view-fn lifecycle-callbacks dom-id initial-state init-fn cfg handler-map]}]
+  [{:keys [cmp-id view-fn lifecycle-callbacks dom-id initial-state init-fn cfg handler-map state-pub-handler]}]
   (let [reagent-cmp-map (merge lifecycle-callbacks {:reagent-render view-fn})
         mk-state (partial init reagent-cmp-map dom-id initial-state init-fn)]
     {:cmp-id            cmp-id
-                          :state-fn          mk-state
-                          :handler-map       handler-map
-                          :state-pub-handler state-pub-handler
-                          :opts              (merge cfg {:watch :local})}))
+     :state-fn          mk-state
+     :handler-map       handler-map
+     :state-pub-handler (or state-pub-handler default-state-pub-handler)
+     :opts              (merge cfg {:watch :local})}))
 
 (defn component
   [cmp]